CNPJ verification when Registering (#3)
A new component "RegisterForm" was created, which contains
the new register form, including the CNPJ field. The submit
button now sends an api request to the backend, which makes
an api request to ReceitaWS, gets information regarding that
CNPJ, and returns it to the frontend. Register is now
temporarily not working.