Enzoabate123 orcamento (#6)
* Fiz uma landing page com uma navbar simples e adicionei alguns componentes que podem ser reutilizados mais tarde

* -a

* fiz a landing page com o banner principal sendo os dois buttons call to action pras respectivas páginas

* Adicionei pagina de orcamento

* alguma merda ai

* Landing Page 0.1 (#2)

* Fiz uma landing page com uma navbar simples e adicionei alguns componentes que podem ser reutilizados mais tarde

* -a

* fiz a landing page com o banner principal sendo os dois buttons call to action pras respectivas páginas

* Upadating gitignore

* Upadating gitignore

* Upadating gitignore

* Upadating gitignore

* CNPJ verification when Registering (#3)

A new component "RegisterForm" was created, which contains
the new register form, including the CNPJ field. The submit
button now sends an api request to the backend, which makes
an api request to ReceitaWS, gets information regarding that
CNPJ, and returns it to the frontend. Register is now
temporarily not working.

* Fixing requirements

* Fixing register page

* Fixing CNPJ validation

Django's default jwt authentication settings were causing the
cnpj validation API to have restricted access even with the
AllowAny option. By adding a new setting above it, the problem
was fixed.

* Company name autocomplete

When registering, whenever a valid CNPJ is typed, the form
automatically fills the Company Name input tag with
the results from the api call.

* Added email field to registration form

---------

Co-authored-by: enzoabate123 <enzoabate12345678910@gmail.com>
Co-authored-by: Enzo Soares Abate <109160184+enzoabate123@users.noreply.github.com>